LPS-196468 Fix search of the LeftSidebar
