+++ modules/apps/object/object-web/types/src/main/resources/META-INF/resources/js/components/ModelBuilder/LeftSidebar/TreeViewComponent.d.ts
@@ -3,15 +3,19 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-/// <reference types="react" />
-
+import React from 'react';
 import './LeftSidebar.scss';
+import {LeftSidebarItem} from '../types';
 export default function TreeViewComponent({
-	query,
-	setEmptySearch,
+	expandedKeys,
+	leftSidebarOtherObjectFoldersItems,
+	leftSidebarSelectedObjectFolderItem,
+	setExpandedKeys,
 	showActions,
 }: {
-	query: string;
-	setEmptySearch: (value: boolean) => void;
+	expandedKeys: Set<React.Key>;
+	leftSidebarOtherObjectFoldersItems: LeftSidebarItem[];
+	leftSidebarSelectedObjectFolderItem: LeftSidebarItem;
+	setExpandedKeys: React.Dispatch<React.SetStateAction<Set<React.Key>>>;
 	showActions?: boolean;
 }): JSX.Element;
